Player can Move and Spin

+++ Vector2.h
@@ -2,8 +2,8 @@
 class Vector2
 {
 public:
-	int x;
-	int y;
-	Vector2(int x, int y);
+	float x;
+	float y;
+	Vector2(float x, float y);
 };
